Fixed random, added startup checker to wait for CSGO to open.

+++ src/main/kotlin/org/abendigo/util/Randoms.kt
@@ -2,7 +2,7 @@ package org.abendigo.util
 
 import java.util.concurrent.ThreadLocalRandom.current
 
-fun random(min: Int, max: Int) = min + current().nextInt() * (max - min)
+fun random(min: Int, max: Int) = (min + current().nextDouble() * (max - min)).toInt()
 
 fun random(max: Int) = random(0, max)
